+++ src/com/Main.java
@@ -1,0 +1,17 @@
+package com;
+
+import java.util.Scanner;
+import com.view.ViewMenu;
+
+public class Main {//Main
+    static int choiceDealerMenu,choiceSellerMenu,choiceMainMenu,choice;
+
+
+    public static void main(String[] args) {
+        Scanner reader = new Scanner(System.in);
+        ViewMenu.Mainmenu(reader);
+
+    }
+
+
+}//Main
