+++ src/com/model/Car.java
@@ -1,8 +1,8 @@
-package com.company;
+package com.model;
 
 public class Car {
     String brand;
-    int numberofdoors;
+    String numberofdoors;
     String color;
     double km;
     String carLicense;
@@ -11,10 +11,14 @@ public class Car {
     //Hello
     //hello
 
+    public String getCarLicense() {
+        return carLicense;
+    }
+
     //Constructor empty
     public Car(){}
 
-    public Car(String brand, int numberofdoors, String color, double km, String carLicense, String insurance, double price){
+    public Car(String brand, String numberofdoors, String color, double km, String carLicense, String insurance, double price){
         this.brand = brand;
         this.numberofdoors = numberofdoors;
         this.color = color;
